requirements.txt
streamlit
psycopg2-binary
pandas
matplotlib
python-dotenv

